requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.0
flask-login==0.6.3
Werkzeug==3.0.0
psycopg2==2.9.6
python-dotenv==1.0.1